pyproject.toml
[project]
name = "cad-quoting-tool"
version = "0.0.0"
description = "Desktop workflow for quoting CAD-driven jobs."
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "requests>=2.31",
    "requests-pkcs12>=1.22",
    "truststore>=0.9.2",
    "python-dotenv>=1.0",
    "pandas>=1.5",
    "openpyxl>=3.1",
    "OCP>=7.7",
    "trimesh>=3.22",
    "ezdxf>=1.0",
    "llama-cpp-python>=0.2.0",
    "pandas-stubs==2.2.2.240514",
    "types-ezdxf==1.0.0",
    "types-pythonocc-core==7.7.0.1",
    "types-PyMuPDF==1.24.3.20240525",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
cad_quoter = [
    "resources/**/*",
    "pricing/resources/**/*",
    "llm/**/*",
]

[tool.ruff]
line-length = 100
fix = true
target-version = "py311"

[tool.ruff.lint]
select = ["F401", "F841", "E", "W", "I"]
ignore = []

[tool.black]
line-length = 100

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
